requirements.txt
fastapi
uvicorn
google-adk
google-cloud-aiplatform
google-cloud-firestore
python-dotenv